Issue #55: Updates template_preprocess_node().
This commit drops a lot of old functionality from the 7.x branch.

We no longer strictly need an "unpublished" class. Partly because D8's
attribute array makes a node--published class available and partly
because we could just use the node object's isPublished() method. I've
left the code for it in place as an example of how it could be done.

The theme hook suggestions were dropped from preprocess_node since D8
actually already has most of the suggestions we were adding in 7.x.
I did create some theme suggestions for the page templates however based
on the entity type. This is done with hook_theme_suggestions_alter().