Fixing some issues that come up out of the gate trying to use windup
Fixed Gruntfile eslint issue, properties should be quoted
Establishing the .estlinrc as root so it doesn't try to use Drupal core's .eslintrc, this is an issue because it requires airbnb standards which requires a lot of other things that we haven't specified in our package.json. This results in build errors.